feat(payment): PAYPAL-4066 shipping address component mapper created
